requirements.txt
Django>=4.2,<5
psycopg2-binary>=2.9
djangorestframework>=3.14.0
django-cors-headers>=3.15.0
requests>=2.28.0